Half precision for embedder and reranker
Implemented halfprecision for reranker and embedder in RAG.
Logic can be used to apply heavier quantization as well - however main memory gain is already done.

Main vector lib creation for chroma is done on full precision but when running model with RAG embedder is loaded in halfprecision so we save memory during iference.